# Optimise DB query - avoid unnecessary join