--- a/babbage/model/attribute.py
+++ b/babbage/model/attribute.py
@@ -18,6 +18,18 @@ class Attribute(Concept):
     def datatype(self):
         return self.spec.get('type')
 
+    @property
+    def column_name(self):
+        #        import pdb; pdb.set_trace()
+        if self.dimension.join_column_name and self == self.dimension.key_attribute:
+            return self.dimension.join_column_name
+        else:
+            return self._column_name
+
+    @column_name.setter
+    def column_name(self, val):
+        self._column_name = val
+
     def __repr__(self):
         return "<Attribute(%s)>" % self.ref